+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   Note: older library versions also support the new languages, this update only
   adds new code constants.
 ### Changed
+* Change return type of `nonRegionalLanguageCode()` to newly-added type
+  `NonRegionalLanguageCode`.
 ### Deprecated
 ### Removed
 ### Fixed